refactor ColourSwatch into a separate file
