+++ docs/components/demo/index.ts
@@ -1,0 +1,2 @@
+export * from './ColourSwatch';
+export { default as ColourSwatch } from './ColourSwatch';
