+++ docs/package.json
@@ -25,6 +25,7 @@
     "build": "parcel build index.html"
   },
   "dependencies": {
+    "components": "^0.1.0",
     "react-app-polyfill": "^1.0.5"
   },
   "devDependencies": {
